fix: support GetByPath and skipFast fallback in arm64
